# docs/openapi.yml
openapi: 3.0.3
info:
  title: LilyOpenCMS API
  description: |
    A comprehensive API for LilyOpenCMS content management system.
    
    ## Features
    - **Content Management**: News, albums, chapters, and media
    - **User Management**: Profiles, authentication, and permissions
    - **Public API**: Multiplatform access for mobile apps and third-party integrations
    - **SEO Management**: Comprehensive SEO tools and analytics
    - **Analytics**: Performance monitoring and user statistics
    
    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Public API endpoints (`/api/public/*`) do not require authentication.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Please implement appropriate retry logic.
    
    ## Support
    For API support, please refer to the documentation or contact the development team.
  version: 1.0.0
  contact:
    name: LilyOpenCMS Development Team
    email: support@lilyopencms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}"
    description: Custom server (set scheme/host below to try any deployment)
    variables:
      scheme:
        enum:
          - http
          - https
        default: https
      host:
        default: localhost:5000
        description: Hostname and optional port (e.g., api.example.com or localhost:5000)
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (external)
      description: JSON registration that creates a pending account (is_active=false, verified=false)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                email: { type: string, nullable: true }
                first_name: { type: string, nullable: true }
                last_name: { type: string, nullable: true }
              required: [username, password]
            examples:
              sample:
                summary: Basic registration
                value:
                  username: demo_user
                  password: demopass123
                  email: demo@example.com
                  first_name: Demo
                  last_name: User
      responses:
        '201':
          description: Registration successful; account pending approval
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user_id: { type: integer }
                  status: { type: string, example: pending_approval }
                example:
                  message: Registration successful. Account pending approval.
                  user_id: 123
                  status: pending_approval
        '400': { description: Validation error }
        '409': { description: Username or email already exists }
      x-codeSamples:
        - lang: cURL
          label: Register
          source: |
            curl -X POST "http://localhost:5000/api/auth/register" \
              -H "Content-Type: application/json" \
              -d '{
                "username": "demo_user",
                "password": "demopass123",
                "email": "demo@example.com"
              }'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (external)
      description: Returns access and refresh tokens for bearer authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
            examples:
              sample:
                summary: Login payload
                value:
                  username: demo_user
                  password: demopass123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  token_type: { type: string, example: bearer }
                  expires_in: { type: integer, example: 3600 }
                  user:
                    type: object
                example:
                  access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type: bearer
                  expires_in: 3600
                  user:
                    id: 123
                    username: demo_user
                    role: GENERAL
                    is_premium: false
                    verified: true
        '401': { description: Invalid credentials }
        '403': { description: Account pending or suspended }
      x-codeSamples:
        - lang: cURL
          label: Login
          source: |
            curl -X POST "http://localhost:5000/api/auth/login" \
              -H "Content-Type: application/json" \
              -d '{
                "username": "demo_user",
                "password": "demopass123"
              }'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
              required: [refresh_token]
            examples:
              sample:
                summary: Refresh payload
                value:
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  token_type: { type: string, example: bearer }
                  expires_in: { type: integer, example: 3600 }
                example:
                  access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type: bearer
                  expires_in: 3600
        '401': { description: Invalid or expired refresh token }
      x-codeSamples:
        - lang: cURL
          label: Refresh tokens
          source: |
            curl -X POST "http://localhost:5000/api/auth/refresh" \
              -H "Content-Type: application/json" \
              -d '{
                "refresh_token": "<REFRESH_TOKEN>"
              }'
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user (external bearer)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  username: { type: string }
                  email: { type: string, nullable: true }
                  role: { type: string, nullable: true }
                  is_premium: { type: boolean }
                  verified: { type: boolean }
                  last_login: { type: string, format: date-time, nullable: true }
                example:
                  id: 123
                  username: demo_user
                  email: demo@example.com
                  role: GENERAL
                  is_premium: false
                  verified: true
                  last_login: 2025-09-05T14:41:00Z
        '401': { description: Unauthorized }
      x-codeSamples:
        - lang: cURL
          label: Get current user
          source: |
            curl "http://localhost:5000/api/auth/me" \
              -H "Authorization: Bearer <ACCESS_TOKEN>"

  # =============================
  # External Profile & Account API
  # =============================
  /api/auth/profile:
    get:
      tags: [Auth]
      summary: Get current user's profile and user fields
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile returned
    put:
      tags: [Auth]
      summary: Update user and profile fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string, nullable: true }
                last_name: { type: string, nullable: true }
                email: { type: string, nullable: true }
                birthdate: { type: string, format: date, nullable: true }
                bio: { type: string, nullable: true }
                pronouns: { type: string, nullable: true }
                short_bio: { type: string, nullable: true }
                location: { type: string, nullable: true }
                website: { type: string, nullable: true }
                social_links: { type: object, additionalProperties: { type: string } }
      responses:
        '200': { description: Profile updated }
  /api/auth/profile/privacy:
    patch:
      tags: [Auth]
      summary: Update privacy flags
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                show_email: { type: boolean }
                show_birthdate: { type: boolean }
      responses:
        '200': { description: Privacy updated }
  /api/auth/profile/username:
    patch:
      tags: [Auth]
      summary: Change username
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                current_password: { type: string }
              required: [username, current_password]
      responses:
        '200': { description: Username changed }
        '400': { description: Validation error }
        '409': { description: Username already taken }
  /api/auth/account/change-password:
    post:
      tags: [Auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password: { type: string }
                new_password: { type: string }
                confirm_password: { type: string }
              required: [current_password, new_password, confirm_password]
      responses:
        '200': { description: Password changed }
        '400': { description: Validation error }
  /api/auth/account:
    delete:
      tags: [Auth]
      summary: Delete account
      security:
        - bearerAuth: []
      responses:
        '200': { description: Account deleted }
        '500': { description: Server error }

  # ADDITIONAL PUBLIC API ENDPOINTS (Newly Implemented)
  # =============================================================================
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via cookies
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external ad access and third-party integrations
security:
  - sessionAuth: []
  - bearerAuth: []